requirements.txt
langgraph
langchain
langchain-openai
langchain-community
python-dotenv
streamlit
requests
tavily-python
finnhub-python
yfinance